HSEARCH-4705 ES test dialect cleanups
- remove the interface and have just a simple class
- use actual version in messages instead of dialect
