+++ util/internal/integrationtest/backend/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/backend/elasticsearch/dialect/ElasticsearchTestDialect.java
@@ -6,38 +6,68 @@
  */
 package org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect;
 
+import static org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect.ElasticsearchVersionUtils.isActualVersionBetweenIncluding;
+import static org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect.ElasticsearchVersionUtils.isActualVersionLess;
+
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
 import org.hibernate.search.backend.elasticsearch.ElasticsearchVersion;
+import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
 
-public interface ElasticsearchTestDialect {
+public class ElasticsearchTestDialect {
 
-	static ElasticsearchTestDialect get() {
-		return new ElasticsearchTestDialectImpl();
+	private static final ElasticsearchVersion ACTUAL_VERSION = ElasticsearchVersion.of(
+			System.getProperty( "org.hibernate.search.integrationtest.backend.elasticsearch.version" )
+	);
+
+	public static ElasticsearchTestDialect get() {
+		return new ElasticsearchTestDialect();
 	}
 
-	static ElasticsearchVersion getActualVersion() {
-		return ElasticsearchVersion.of( System.getProperty( "org.hibernate.search.integrationtest.backend.elasticsearch.version" ) );
+	public static ElasticsearchVersion getActualVersion() {
+		return ACTUAL_VERSION;
 	}
 
-	boolean isEmptyMappingPossible();
 
-	Optional<URLEncodedString> getTypeNameForMappingAndBulkApi();
+	public boolean isEmptyMappingPossible() {
+		return isActualVersionLess( ACTUAL_VERSION, "elastic:6.9.0" );
+	}
 
-	Boolean getIncludeTypeNameParameterForMappingApi();
+	@SuppressWarnings("deprecation")
+	public Optional<URLEncodedString> getTypeNameForMappingAndBulkApi() {
+		if ( isActualVersionLess( ACTUAL_VERSION, "elastic:6.9.0" ) ) {
+			return Optional.of( Paths.DOC );
+		}
+		return Optional.empty();
+	}
 
-	List<String> getAllLocalDateDefaultMappingFormats();
+	public Boolean getIncludeTypeNameParameterForMappingApi() {
+		if ( isActualVersionBetweenIncluding( ACTUAL_VERSION, "elastic:6.4.0", "elastic:6.9.0" ) ) {
+			return Boolean.TRUE;
+		}
+		return null;
+	}
 
-	default String getFirstLocalDateDefaultMappingFormat() {
+	public List<String> getAllLocalDateDefaultMappingFormats() {
+		if ( isActualVersionLess( ACTUAL_VERSION, "elastic:6.9.0" ) ) {
+			return Arrays.asList( "yyyy-MM-dd", "yyyyyyyyy-MM-dd" );
+		}
+		return Collections.singletonList( "uuuu-MM-dd" );
+	}
+
+	public boolean supportsIsWriteIndex() {
+		return !isActualVersionLess( ACTUAL_VERSION, "elastic:6.4.0" );
+	}
+
+	public String getFirstLocalDateDefaultMappingFormat() {
 		return getAllLocalDateDefaultMappingFormats().get( 0 );
 	}
 
-	default String getConcatenatedLocalDateDefaultMappingFormats() {
+	public String getConcatenatedLocalDateDefaultMappingFormats() {
 		return String.join( "||", getAllLocalDateDefaultMappingFormats() );
 	}
-
-	boolean supportsIsWriteIndex();
-
 }
